+++ epictrack-web/src/components/myWorkplans/CardList.tsx
@@ -12,14 +12,7 @@ const CardList = () => {
   }
 
   return (
-    <Grid
-      container
-      direction="row"
-      alignItems="flex-start"
-      justifyContent="space-evenly"
-      spacing={2}
-      columns={{ xs: 12 }}
-    >
+    <Grid container direction="row" spacing={2} columns={{ xs: 12 }}>
       {workplans.map((workplan) => {
         return (
           <Grid item xs={4}>
